Arch

+++ myportfy-backend/src/main/java/com/myportfy/services/serviceImpl/UserServiceImpl.java
@@ -4,16 +4,18 @@ import com.myportfy.domain.User;
 import com.myportfy.domain.enums.Role;
 import com.myportfy.dto.PasswordDto;
 import com.myportfy.dto.UserPrincipal;
+import com.myportfy.dto.user.UserGetDto;
 import com.myportfy.repositories.PostRepository;
 import com.myportfy.repositories.UserRepository;
 import com.myportfy.services.IEmailService;
-import com.myportfy.services.IImageService;
 import com.myportfy.services.IS3Service;
 import com.myportfy.services.IUserService;
 import com.myportfy.services.exceptions.AuthorizationException;
 import com.myportfy.services.exceptions.ObjectNotFoundException;
 import com.myportfy.utils.FillNullProperty;
+import com.myportfy.utils.image.ImageUtils;
 import lombok.extern.slf4j.Slf4j;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
@@ -30,6 +32,7 @@ import java.awt.image.BufferedImage;
 import java.net.URI;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
@@ -43,6 +46,7 @@ public class UserServiceImpl implements IUserService {
 
     private final String USER_NOT_FOUND_MESSAGE = "Não encontrei nenhum usuário.. 😫";
     private final String AUTHORIZARTION_EXCEPTION_MESSAGE = "Acesso não autorizado.";
+    private final String DEFAULT_PICTURE = "https://my-portfy.s3.amazonaws.com/People.svg";
 
     @Autowired
     private UserRepository userRepository;
@@ -52,16 +56,21 @@ public class UserServiceImpl implements IUserService {
     private BCryptPasswordEncoder bCryptPasswordEncoder;
     @Autowired
     private IS3Service s3Service;
-    @Autowired
-    private IImageService imageService;
     @Autowired @Lazy
     private IEmailService emailService;
+    @Autowired
+    private ModelMapper modelMapper;
 
     @Override
     @Transactional(readOnly = true)
-    public Page<User> findAll(Pageable pageable) {
+    public Page<UserGetDto> findAll(Pageable pageable) {
         log.info("Fetching all users");
-        return userRepository.findAll(pageable);
+        return userRepository.findAll(pageable).map(x -> modelMapper.map(x, UserGetDto.class));
+    }
+
+    @Override
+    public Page<?> findAllGeneric(Pageable pageable) {
+        return null;
     }
 
     @Override
@@ -81,15 +90,18 @@ public class UserServiceImpl implements IUserService {
         object.setId(null);
         object.setPassword(bCryptPasswordEncoder.encode(object.getPassword()));
         object.setCreatedAt(now());
-        object.setProfilePictureURL("https://my-portfy.s3.amazonaws.com/People.svg");
+        object.setProfilePictureURL(DEFAULT_PICTURE);
         userRepository.saveAndFlush(object);
         log.info("New user created: {}", object.getUsername());
         emailService.sendAccountConfirmation(object);
     }
 
     @Override
+    public void create(User object, Object arg) {}
+
+    @Override
     @Transactional(propagation = REQUIRED)
-    public void update(User object) {
+    public User update(User object) {
         UserPrincipal user = currentUserLoggedIn();
         if (!user.hasRole(ADMIN) && !object.getId().equals(user.getId())) {
             log.error("Authorization exception for user {} on update user {}", user.getUsername(), object.getUsername());
@@ -112,8 +124,13 @@ public class UserServiceImpl implements IUserService {
         updateObject.setCreatedAt(createAt);
         updateObject.setUpdatedAt(now());
         updateObject.setRoles(role);
-        userRepository.save(updateObject);
         log.info("User updated, id: {}", object.getId());
+        return userRepository.save(updateObject);
+    }
+
+    @Override
+    public User update(User object, Object arg) {
+        return null;
     }
 
     @Override
@@ -217,10 +234,10 @@ public class UserServiceImpl implements IUserService {
     @Transactional(propagation = REQUIRED)
     public void uploadProfilePicture(BufferedImage jpgImage, String fileName, User user) {
         URI uri = s3Service.uploadFile(
-                imageService.getInputStream(jpgImage, "JPG"),
+                ImageUtils.getInputStream(jpgImage, "JPG"),
                 fileName,
                 "image");
-        if (user.getProfilePictureURL() != null) {
+        if (user.getProfilePictureURL() != null && !Objects.equals(user.getProfilePictureURL(), DEFAULT_PICTURE)) {
             deleteProfilePicture(user);
         }
         user.setProfilePictureURL(uri.toString());
